Create main readme for viperleed project
@fkraushofer, @amimre, @schmid-iap: Feel free to add more details.

This is something that Lutz suggested to add (see https://github.com/viperleed/viperleed/issues/129) so people have a little bit of info when they land on the page till we make it public.

Also, @schmid-iap got a "how to cite" request from Phil Woodruff (for a mini review) and I suggested he refers to the github for now. @schmid-iap suggested to have some text before using this as a reference.